adjustemtns
